Continued reworking the code to make it simplier and easier to reuse.
Created a patterns class which can be used to store various types of patterns/filters for spidering. Initially this was used to filter only on images. By pulling this out, we'll be able to do other types by updating the filters and calling them as well.

Create a BinarySizes, which can be used for minimum binary size spidering.