Finished reworking the code to make it more flexible for different spider search patterns.
Crawler4j is coded in a way that it references static patterns. Because of this we have to create new classes to specify different types of patterns. (Since users should be able to run searches in parallel.)

So users can search in parallel, as long as the patterns are the same. CrawlerImagesOnly* is an example of Crawler is extended to allow for things to work. Currently they are the same patterns, however these can be changed, e.g. So can download movie files and pdfs too.