Started work on getting uploaded blobs to db.
Found an example that wasn't fully complete, however I updated it so that it compiles now.

I referenced code from the first project of the course, for the mysql connection info. I pulled it over as an example, however changed the project so that it references the connection jar through the maven pom instead of being explicity added.