Checkin for temp progress made so far related to blob work.
This is not working yet, however I've updated some references to get path locations as desired by our spider portion of the program.

I've also simplified some of the example code so that we only deal with one binary... this will need to be updated to a real blob field refernce once things are working.